Rename .kxio just to .io
