+++ formats/json-io/commonMain/src/kotlinx/serialization/json/io/internal/IoJsonStreams.kt
@@ -2,14 +2,14 @@
  * Copyright 2017-2022 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package kotlinx.serialization.json.kxio.internal
+package kotlinx.serialization.json.io.internal
 
 import kotlinx.io.*
 import kotlinx.serialization.json.internal.*
 
 private const val QUOTE_CODE = '"'.code
 
-internal class JsonToKxioStreamWriter(private val sink: Sink) : InternalJsonWriter {
+internal class JsonToIoStreamWriter(private val sink: Sink) : InternalJsonWriter {
 
     override fun writeLong(value: Long) {
         write(value.toString())
@@ -34,7 +34,7 @@ internal class JsonToKxioStreamWriter(private val sink: Sink) : InternalJsonWrit
     }
 }
 
-internal class KxioSerialReader(private val source: Source): InternalJsonReaderCodePointImpl() {
+internal class IoSerialReader(private val source: Source): InternalJsonReaderCodePointImpl() {
     override fun exhausted(): Boolean = source.exhausted()
     override fun nextCodePoint(): Int = source.readCodePointValue()
 }
